spring-25-hw / overall refactoring
