+++ spring-25-hw/build.gradle
@@ -29,11 +29,11 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter'
     implementation 'org.springframework.boot:spring-boot-starter-batch'
     implementation 'org.springframework.shell:spring-shell-starter:2.0.1.RELEASE'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
+    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
 
     implementation 'com.h2database:h2'
     implementation 'org.liquibase:liquibase-core'
