+++ spring-25-hw/src/main/java/ru/skubatko/dev/otus/spring/hw25/dto/BookDto.java
@@ -27,12 +27,6 @@ public class BookDto {
 
     private String comments;
 
-    public BookDto(String name, String author, String genre) {
-        this.name = name;
-        this.author = author;
-        this.genre = genre;
-    }
-
     public static BookDto toDto(SqlBook book) {
         if (book == null) {
             return null;
