+++ spring-25-hw/src/main/java/ru/skubatko/dev/otus/spring/hw25/model/nosql/NoSqlBook.java
@@ -4,14 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Data
-@Document
+@Document(collection = "books")
 @NoArgsConstructor
 @AllArgsConstructor
 public class NoSqlBook {
@@ -21,8 +20,6 @@ public class NoSqlBook {
     private String name;
     private String author;
     private String genre;
-
-    @DBRef
     private List<NoSqlComment> comments = new ArrayList<>();
 
     public NoSqlBook(String name, String author, String genre, List<NoSqlComment> comments) {
