+++ spring-25-hw/src/main/java/ru/skubatko/dev/otus/spring/hw25/model/sql/SqlComment.java
@@ -33,9 +33,4 @@ public class SqlComment {
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "book_id", nullable = false)
     private SqlBook book;
-
-    public SqlComment(SqlBook book, String commentContent) {
-        this.book = book;
-        this.content = commentContent;
-    }
 }
