+++ spring-25-hw/src/main/java/ru/skubatko/dev/otus/spring/hw25/repository/NoSqlBookRepository.java
@@ -1,0 +1,8 @@
+package ru.skubatko.dev.otus.spring.hw25.repository;
+
+import ru.skubatko.dev.otus.spring.hw25.model.nosql.NoSqlBook;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface NoSqlBookRepository extends MongoRepository<NoSqlBook, String> {
+}
