+++ spring-25-hw/src/main/java/ru/skubatko/dev/otus/spring/hw25/repository/SqlBookRepository.java
@@ -1,8 +1,6 @@
-package ru.skubatko.dev.otus.spring.hw25.repository.sql;
+package ru.skubatko.dev.otus.spring.hw25.repository;
 
-import ru.skubatko.dev.otus.spring.hw25.model.sql.SqlAuthor;
 import ru.skubatko.dev.otus.spring.hw25.model.sql.SqlBook;
-import ru.skubatko.dev.otus.spring.hw25.model.sql.SqlGenre;
 
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -14,10 +12,4 @@ public interface SqlBookRepository extends JpaRepository<SqlBook, Long> {
     @EntityGraph(attributePaths = {"author", "genre", "comments"})
     @Override
     List<SqlBook> findAll();
-
-    SqlBook findByName(String name);
-
-    List<SqlBook> findByAuthor(SqlAuthor author);
-
-    List<SqlBook> findByGenre(SqlGenre genre);
 }
